Option to source a file on the host
When executing a command through ssh, the env variables of the user on the
host are not used, and a basic set of variables is used instead.
The startup-script option allows to pass the path of a script that will be
sourced before each ssh command.
More simply we had "source SCRIPT_PATH &&" before each command when the
property is set.