Workaround proxy issues with rpm
When the host is behind a proxy, the rpm -Uvh command fails.
Rpm uses curl behind the scene but it ignores the proxy settings in the env
variables.
We now fetch the rpm file using curl and store it in a temporary location
before executing rpm -Uvh against that it.
We need to use a temporary file because rpm doesn't work with stdin..