Add --run-list and --json-attributes options to generate node config
Sometimes it is more convenient to specify the run list on command line
option than create or edit the JSON file. Someone may even want to
set other node attributes.

This commit adds `--run-list` and `--json-attributes` options to
`knife prepare`. The options and their format is the same than in
for example `knife bootstrap`.