Fix `--omnibus-options` passing to Omnibus Installer
It seems that all the refactoring in PR #99 skipped renaming the option
in `KnifeSolo::Bootstraps`.

There are no tests for the Omnibus options and they are a bit tricky to
implement, so they are left as an exercise for the reader. Maybe
something like #assert_chef_solo_option in the Cook tests?