Rename Cook command's `--skip-chef-check` option to `--no-chef-check`
Knife convention is to use "--no-option" or "--[no-]option". In that
case the "Mixlib::CLI" class allows the option with or without the "no".
And this gives also the opportunity to set the default on knife
configuration and then override it on command line.

The old "skip" option is left with a deprecation warning for some period
of time to maintain backwards compatibility.