Remove kitchen validation
Now that we no longer have a solo.rb none of the typical kitchen files are actually required. Even a node config will get generated at runtime if it doesn't already exist.

This test was originally included as a guard against syncing more than was intentional. I'm unable to come up with a way to effectively do that without getting tricky. Ideas are welcome.