Fix :librarian configuration handling
Librarian-Chef was not run by default when SoloCook/Bootstrap was
invoked by for example `knife bootstrap --solo` (from #207) or `knife
digital_ocean droplet create --solo` (from knife-digital_ocean).

The reason is that the `Knife.config` is not loaded by default values
when the Knife class is initialized. While this could be fixed in
the invoking code it is much cleaner to do moving the default value
setting to later in the code.

As an extra bonus, by using our `config_value` helper we also make it
possible to use `knife[:librarian]` configuration option with Chef 10,
too.