Remove most version detections from EL family distros
Omnibus installer supports generally new releases fast, so we
don't need to check the version. There is not much difference
if we don't support some version or if the omnibus installer
fails. In both cases the user has to install Chef manually
(or file an issue to us to make an exception).