Upload only the needed node config and make the path configurable
Use `node_path` to specify the directory where node configuration json
files are stored. Default to "nodes" as earlier. It is worth noting that
`node_path` defaults to "/var/chef/nodes" on Chef versions before 11.6.0.
But as that directory should not exist, we will fallback to "nodes".

We now only upload the json for the cooked node.