Explicitly set root path for default `*_path` options
Chef 11.8.0 (with mixlib-config 2.0) sets `*_path` options by default
based on `cookbook_path`. If it is an array (as always in knife-solo),
the other paths are also arrays. And #300 will catch that and error out.

While the conversion to Array is a regression in Chef itself, we anyway
want to set the default path relative to our kitchen root, as
`cookbook_path` components might be pointing to somewhere else.