Verify that `node_path` is a String
`node_path` made it back to Chef's default configuration in 11.8.0. And
as pre 0.3.0 setups won't normally have it explicitly configured, it
defaults to being an Array.

Also a user might set it to be an Array, which leads knife-solo crashing
with the infamous `ERROR: TypeError: no implicit conversion of Array
into String`. This escaped from us in GH-300.