Include remote user and port in ssh control path
Only one user can cook a node from a single workstation when ssh
multiplexing is enabled, as it is by default, but the same control
socket is shared for ssh sessions with different users and ports.

This causes problems when bootstrapping new nodes if root logins are
disabled and/or the sshd port is changed during the initial chef run.

Fixed by using the remote user and port in the ssh control path.

Fixes #444