Shorten ssh control path
My previous commit to "fix" the ssh control path by including the remote
user and port also made it more likely that the resulting path would
exceed the maximum allowed length of the path to a unix domain socket.

I've replaced the use of '%r@%h:%p' in the ssh control path with '%C',
which is substituted by a hash of the concatenation of '%l%h%p%r', so it
still includes the remote user, host and port, but is also a predictable
length and less likely to exceed the maximum allowed length of a path to
a unix domain socket (104 "chars" on OSX, 108 I think on Linux).

The change was made after I ran into a "too long for Unix domain socket"
error myself when cooking a node, and has been tested in production.